+++ app.js
@@ -7,8 +7,8 @@ const loginRouter = require('./routes/login');
 const dashboardRouter = require('./routes/dashboard');
 const settingsRouter = require('./routes/settings');
 const teachersRouter = require('./routes/teachers');
-const schoolsRouter = require('./routes/schools');
-const schoolDistrictsRouter = require('./routes/schoolDistricts');
+const sites = require('./routes/sites');
+const areas = require('./routes/areas');
 const students = require('./routes/students');
 
 const app = express();
@@ -29,8 +29,8 @@ app.use('/dashboard', dashboardRouter);
 app.use('/login', loginRouter);
 app.use('/settings', settingsRouter);
 app.use('/teachers', teachersRouter);
-app.use('/schools', schoolsRouter);
-app.use('/schoolDistricts', schoolDistrictsRouter);
+app.use('/sites', sites);
+app.use('/areas', areas);
 app.use('/students', students);
 // school district
 // schools
